+++ common/include/Registry.h
