+++ common/src/Registry.cpp
